Add 64-bit support

+++ src/SBLib/include/network.h
@@ -2,7 +2,7 @@
 
 #define _WINSOCK2API_   /* Prevent inclusion of winsock2.h in enet.h */
 typedef struct fd_set fd_set;
-typedef unsigned int SOCKET;
+typedef ptrdiff_t SOCKET;
 #include <enet/enet.h>
 
 typedef DWORD DPID;
